fix: desktop history details refresh on record delete

+++ lib/desktop_screens/desktop_history/widgets/desktop_received_file_list_tile.dart
@@ -679,7 +679,9 @@ class _DesktopReceivedFilesListTileState
 
                   if (await CommonUtilityFunctions().isFilePresent(filePath)) {
                     var file = File(filePath);
-                    file.deleteSync();
+                    if (await file.existsSync()) {
+                      file.deleteSync();
+                    }
                   }
                 });
